banner, bg and tabs completed
