+++ public/index.html
@@ -16,7 +16,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Проектная кухня Практикума</title>
+    <title>Pepelats-Q Community</title>
   </head>
   <body>
     <div id="root"></div>
