+++ src/components/ArticlePreview/ArticlePreview.js
@@ -4,6 +4,7 @@ import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import agent from '../../agent';
 import { ARTICLE_FAVORITED, ARTICLE_UNFAVORITED } from '../../constants/actionTypes';
+import styles from './style.module.css'
 
 const FAVORITED_CLASS = 'btn btn-sm btn-primary';
 const NOT_FAVORITED_CLASS = 'btn btn-sm btn-outline-primary';
@@ -32,7 +33,7 @@ const ArticlePreview = ({ article, favorite, unfavorite }) => {
   };
 
   return (
-    <div className='article-preview'>
+    <div className={styles.article_preview}>
       <div className='article-meta'>
         <Link to={`/@${article.author.username}`}>
           <img src={article.author.image} alt={article.author.username} />
