+++ src/components/ArticlePreview/style.module.css
@@ -1,0 +1,4 @@
+.article_preview {
+    border-top: 1px solid #2F2F37;
+    padding: 1.5rem 0;
+}
