+++ src/components/Header/header.module.css
@@ -25,6 +25,8 @@
   text-decoration: none;
   display: flex;
   align-items: center;
+  transition: color 0.4s ease-in;
+
 }
 
 .header {
@@ -53,11 +55,13 @@
   color: #f2f2f3;
   text-decoration: none;
   transition: color 0.4s ease-in;
+
 }
 
 .navLink:hover>.menuIcon {
   filter: invert(97%) sepia(16%) saturate(8%) hue-rotate(181deg) brightness(99%) contrast(93%);
   transition: filter 0.4s ease-in;
+
 }
 
 .navLink:hover>.profileIcon {
@@ -102,6 +106,7 @@
   filter: invert(54%) sepia(12%) saturate(778%) hue-rotate(201deg) brightness(97%) contrast(96%);
   padding: 0;
   margin: 0 8px 0 0;
+  transition: filter 0.4s ease-in;
 }
 
 .profileIcon {
