--- src/components/Pages/Home/Banner.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-const Banner = ({ appName, token }) => {
-  if (token) {
-    return null;
-  }
-  return (
-    <div className="banner">
-      <div className="container">
-        <h1 className="logo-font">
-          {appName.toLowerCase()}
-        </h1>
-        <p>Your community project starter pack.</p>
-      </div>
-    </div>
-  );
-};
-
-export default Banner;
