+++ src/components/Pages/Home/Banner/style.module.css
@@ -1,0 +1,45 @@
+.banner {
+    display: flex;
+    height: 164px;
+    justify-content: center;
+    align-items: center;
+    font-family: 'Exo 2', sans-serif;
+
+    text-align: center;
+    box-shadow: inset 0px -1px 0px #2F2F37;
+}
+
+.container {
+    margin-left:auto;
+    margin-right:auto;
+    padding-left:15px;
+    padding-right:15px
+}
+
+.title {
+    color: var(--text-white);
+    font-size: 64px;
+    font-weight: bold;
+    text-shadow: 0 4px 32px var(--accent-color), 0 0 8px var(--accent-color);
+}
+
+.text {
+    color: var(--text-white);
+    font-size: 24px;
+}
+
+@media(min-width:544px) {
+    .container{max-width:576px}
+}
+
+@media(min-width:768px) {
+    .container{max-width:720px}
+}
+
+@media(min-width:992px) {
+    .container{max-width:940px}
+}
+
+@media(min-width:1200px) {
+    .container{max-width:1140px}
+}
