+++ src/components/Pages/Home/index.module.css
@@ -1,0 +1,30 @@
+.home_page {
+    min-height: calc(100vh - 60px);
+    background-color: var(--bg-color);
+}
+
+.container {
+    padding-top: 32px;
+    margin-left:auto;
+    margin-right:auto;
+    padding-left:15px;
+    padding-right:15px;
+}
+
+@media(min-width:544px) {
+    .container{max-width:576px}
+}
+
+@media(min-width:768px) {
+    .container{max-width:720px}
+}
+
+@media(min-width:992px) {
+    .container{max-width:940px}
+}
+
+@media(min-width:1200px) {
+    .container {
+        max-width: 1140px
+    }
+}
