--- src/pages/Home/ToDelete.css
@@ -1,7 +1,0 @@
-.to-delete__title {
-  margin: 40px 0;
-}
-
-.to-delete__list {
-  padding-left: 20px;
-}
